refactor UAV to be abstract class
