+++ controls/sae_2025_ws/src/uav/uav/ModeManager.py
@@ -2,7 +2,7 @@
 import rclpy
 from rclpy.node import Node
 from time import time
-from uav import UAV
+from uav import VTOL, Multicopter
 from uav.autonomous_modes import Mode, LandingMode
 import yaml
 import importlib
@@ -12,11 +12,11 @@ from px4_msgs.msg import VehicleStatus
 
 VISION_NODE_PATH = 'uav.vision_nodes'
 
-class ModeManager(Node):    
+class ModeManager(Node):
     """
     A ROS 2 node for managing UAV modes and mission logic.
     """
-    def __init__(self, mode_map: str, vision_nodes: str, camera_offsets, DEBUG=False, servo_only=False) -> None:
+    def __init__(self, mode_map: str, vision_nodes: str, camera_offsets, DEBUG=False, servo_only=False, is_vtol=False) -> None:
         super().__init__('mission_node')
         self.timer = self.create_timer(0.1, self.spin_once)
         self.modes = {}
@@ -24,7 +24,13 @@ class ModeManager(Node):
         self.active_mode = None
         self.last_update_time = time()
         self.start_time = self.last_update_time
-        self.uav = UAV(self, DEBUG=DEBUG, camera_offsets=camera_offsets)
+
+        # Instantiate appropriate UAV subclass based on vehicle type
+        if is_vtol:
+            self.uav = VTOL(self, DEBUG=DEBUG, camera_offsets=camera_offsets)
+        else:
+            self.uav = Multicopter(self, DEBUG=DEBUG, camera_offsets=camera_offsets)
+
         self.get_logger().info("Mission Node has started!")
         self.setup_vision(vision_nodes)
         self.setup_modes(mode_map)
@@ -271,23 +277,3 @@ class ModeManager(Node):
             data = yaml.safe_load(file)
         return data
     
-
-if __name__ == '__main__':
-    mission_node = ModeManager('test_mode_manager')
-    print(mission_node.modes)
-    print(mission_node.transitions)
-    print(mission_node.active_mode)
-    # mission_node.spin()
-    
-    mission_node.transition('continue')
-    print('transition continue')
-    print(mission_node.modes)
-    print(mission_node.transitions)
-    print(mission_node.active_mode)
-    
-    mission_node.transition('continue')
-    print('transition continu 2')
-    print(mission_node.modes)
-    print(mission_node.transitions)
-    print(mission_node.active_mode)
-
