+++ controls/sae_2025_ws/src/uav/uav/VTOL.py
@@ -1,0 +1,171 @@
+from rclpy.node import Node
+from px4_msgs.msg import VtolVehicleStatus, VehicleCommand
+from rclpy.qos import QoSProfile, QoSReliabilityPolicy, QoSHistoryPolicy, QoSDurabilityPolicy
+import numpy as np
+from uav.UAV import UAV
+
+
+class VTOL(UAV):
+    """
+    VTOL UAV implementation with MC/FW mode switching.
+    Handles both multicopter and fixed-wing flight modes.
+    """
+
+    def __init__(self, node: Node, takeoff_amount=5.0, DEBUG=False, camera_offsets=[0, 0, 0]):
+        # Initialize VTOL-specific attributes before calling super().__init__
+        self.vehicle_type = None  # 'MC' or 'FW' from VtolVehicleStatus
+        self.vtol_vehicle_status = None
+
+        super().__init__(node, takeoff_amount, DEBUG, camera_offsets)
+
+    @property
+    def is_vtol(self) -> bool:
+        """VTOL aircraft can transition between MC and FW modes."""
+        return True
+
+    def vtol_transition_to(self, vtol_state, immediate=False):
+        """
+        Command a VTOL transition.
+        Following https://mavlink.io/en/messages/common.html#MAV_CMD_DO_VTOL_TRANSITION
+
+        Args:
+            vtol_state (str): The desired VTOL state ('MC' or 'FW').
+            immediate (bool): If True, the transition should be immediate.
+        """
+        assert vtol_state in ['MC', 'FW'], "VTOL state must be 'MC' or 'FW'."
+        state = VtolVehicleStatus.VEHICLE_VTOL_STATE_MC if vtol_state == 'MC' else VtolVehicleStatus.VEHICLE_VTOL_STATE_FW
+        immediate = 1 if immediate else 0
+        self._send_vehicle_command(VehicleCommand.VEHICLE_CMD_DO_VTOL_TRANSITION, params={'param1': float(state), 'param2': float(immediate)})
+        self.node.get_logger().info(f"VTOL transition command sent: {state}. Transitioning to {vtol_state} mode.")
+
+    def _calculate_velocity(self, target_pos: tuple, lock_yaw: bool) -> list:
+        """
+        Calculate velocity switching between FW (forward velocity) and MC (proportional) modes.
+
+        Args:
+            target_pos: (x, y, z) target position in local frame
+            lock_yaw: Whether yaw is locked (hovering)
+
+        Returns:
+            [vx, vy, vz] velocity list in m/s
+        """
+        is_fw_mode = (self.vehicle_type == 'FW')
+
+        if is_fw_mode:
+            # Fixed-wing mode: always maintain forward velocity for lift
+            return self._get_fw_forward_velocity()
+        else:
+            # Multicopter mode: use proportional control (same as Multicopter class)
+            if self.local_position is None:
+                # Fallback during startup
+                direction_est = np.array(target_pos)
+                dist_est = np.linalg.norm(direction_est)
+                if dist_est > 0.01:
+                    direction = direction_est / dist_est
+                    return self._calculate_proportional_velocity(direction, dist_est)
+                else:
+                    return [0.0, 0.0, 0.0]
+
+            # Normal case: calculate from current position
+            dx = target_pos[0] - self.local_position.x
+            dy = target_pos[1] - self.local_position.y
+            dz = target_pos[2] - self.local_position.z
+            dist = np.sqrt(dx**2 + dy**2 + dz**2)
+
+            if dist > 0.01:
+                direction = np.array([dx, dy, dz]) / dist
+                return self._calculate_proportional_velocity(direction, dist)
+            else:
+                # At target: hover
+                return [0.0, 0.0, 0.0]
+
+    def _get_fw_forward_velocity(self):
+        """
+        Get forward velocity vector for FW mode to maintain lift.
+
+        Returns:
+            list: [vx, vy, vz] velocity vector in m/s
+        """
+        if self.yaw is not None:
+            return [float(np.cos(self.yaw) * self.default_velocity),
+                    float(np.sin(self.yaw) * self.default_velocity),
+                    0.0]
+        # Last resort: forward in X direction
+        self.node.get_logger().warn("FW mode: Using fallback velocity (X-axis forward) - yaw not available")
+        return [float(self.default_velocity), 0.0, 0.0]
+
+    def _calculate_proportional_velocity(self, direction: np.ndarray, distance: float) -> list:
+        """
+        Calculate velocity using proportional control to prevent oscillation.
+        Used in MC mode. Same implementation as Multicopter class.
+
+        Args:
+            direction (np.ndarray): Unit direction vector [dx, dy, dz]
+            distance (float): Distance to target in meters
+
+        Returns:
+            list: [vx, vy, vz] velocity vector in m/s
+        """
+        # Proportional control with lenient thresholds to reduce oscillation
+        # Full speed above 10m, proportional between 10m-2m, slow below 2m
+        if distance > 10.0:
+            target_speed = self.default_velocity
+        elif distance > 2.0:
+            # Smooth deceleration from 10m to 2m
+            # At 10m: 5 m/s, at 2m: 1 m/s
+            target_speed = max(1.0, self.default_velocity * (distance / 10.0))
+        elif distance > 0.1:
+            # Close to target: gentle approach to prevent overshoot
+            target_speed = 0.8
+        else:
+            # Very close: hover
+            return [0.0, 0.0, 0.0]
+
+        return [float(direction[0] * target_speed),
+                float(direction[1] * target_speed),
+                float(direction[2] * target_speed)]
+
+    def _vtol_vehicle_status_callback(self, msg: VtolVehicleStatus):
+        """
+        Callback for VTOL vehicle status updates.
+        Updates vehicle_type based on the actual VTOL state.
+        """
+        self.vtol_vehicle_status = msg
+        if msg.vehicle_vtol_state == VtolVehicleStatus.VEHICLE_VTOL_STATE_MC:
+            self.vehicle_type = 'MC'
+        elif msg.vehicle_vtol_state == VtolVehicleStatus.VEHICLE_VTOL_STATE_FW:
+            self.vehicle_type = 'FW'
+        # During transitions, maintain the current state (don't change vehicle_type)
+        # VEHICLE_VTOL_STATE_TRANSITION_TO_FW = 1 (still in MC mode)
+        # VEHICLE_VTOL_STATE_TRANSITION_TO_MC = 2 (still in FW mode)
+        elif msg.vehicle_vtol_state == VtolVehicleStatus.VEHICLE_VTOL_STATE_TRANSITION_TO_FW:
+            # Transitioning to FW, but still in MC mode
+            if self.vehicle_type is None:
+                self.vehicle_type = 'MC'
+        elif msg.vehicle_vtol_state == VtolVehicleStatus.VEHICLE_VTOL_STATE_TRANSITION_TO_MC:
+            # Transitioning to MC, but still in FW mode
+            if self.vehicle_type is None:
+                self.vehicle_type = 'FW'
+
+    def _initialize_publishers_and_subscribers(self):
+        """
+        Initialize ROS 2 publishers and subscribers.
+        Extends parent to add VTOL-specific subscriber.
+        """
+        # Call parent to initialize all shared publishers/subscribers
+        super()._initialize_publishers_and_subscribers()
+
+        # Add VTOL-specific subscriber for VtolVehicleStatus
+        qos_profile = QoSProfile(
+            reliability=QoSReliabilityPolicy.BEST_EFFORT,
+            durability=QoSDurabilityPolicy.TRANSIENT_LOCAL,
+            history=QoSHistoryPolicy.KEEP_LAST,
+            depth=1
+        )
+
+        self.vtol_vehicle_status_sub = self.node.create_subscription(
+            VtolVehicleStatus,
+            '/fmu/out/vtol_vehicle_status',
+            self._vtol_vehicle_status_callback,
+            qos_profile
+        )
