+++ controls/sae_2025_ws/src/uav/uav/__init__.py
@@ -1,2 +1,6 @@
 # uav/__init__.py
-from .UAV import UAV
+from .UAV import UAV  # Abstract base class
+from .VTOL import VTOL  # Concrete VTOL implementation
+from .Multicopter import Multicopter  # Concrete multicopter implementation
+
+__all__ = ['UAV', 'VTOL', 'Multicopter']
